Implementación completa seguridad

+++ src/main/java/com/example/sistematransmilenio/service/BusService.java
@@ -2,6 +2,8 @@ package com.example.sistematransmilenio.service;
 
 import com.example.sistematransmilenio.model.Bus;
 import com.example.sistematransmilenio.model.Conductor;
+import com.example.sistematransmilenio.model.Horario;
+import com.example.sistematransmilenio.repository.HorarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.example.sistematransmilenio.repository.BusRepository;
@@ -10,6 +12,9 @@ import java.util.List;
 
 @Service
 public class BusService {
+
+    @Autowired
+    private HorarioRepository horarioRepository;
 
     @Autowired
     private BusRepository busRepository;
@@ -37,13 +42,29 @@ public class BusService {
     }
 
     public boolean eliminarBus(long id ){
-        try{
-            busRepository.deleteById(id);
-        }catch(Exception e){
-            return false;
+
+        if(this.validarBus(id)){
+            try{
+                busRepository.deleteById(id);
+                return true;
+            }catch(Exception e){
+                return false;
+            }
         }
 
         return true;
+    }
+    private boolean validarBus(Long id) {
+        List<Horario> horarios = horarioRepository.getBusByConductorId(id);
+        if(horarios.isEmpty()){
+            return true;
+        }else{
+            for(Horario h: horarios){
+                System.out.println(h.toString());
+            }
+
+            return false;
+        }
     }
 
     public Bus findBusByPlaca(String placa) {
